Implement streaming transcription with Deepgram
The `minigram` program will start `miniogg` and stream its audio
to the Deepgram WebSocket API, printing its events as JSON lines
to standard output.

To achieve low latency, which Opus does very well, we have to lower
two delay settings that `libopusenc` by default sets to high values.

The "muxing delay" defaults to 1 second, and the "decision delay"
defaults to 2 seconds; we now set both to 100 milliseconds.

There are two implementations of the `minigram` program.  One is a
simple bash program that just pipes `miniogg` to `websocat`, and the
other is a JavaScript program for the Bun runtime.  Both do the same
thing.